pyproject.toml
[project]
name = "semantic-cache"
version = "0.1.0"
description = "Enterprise-grade async semantic caching for AI agents using RedisVL"
authors = [
    {name = "Vertector", email = "info@vertector.com"}
]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

keywords = ["redis", "cache", "semantic", "llm", "ai", "async", "redisvl"]

classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "redisvl>=0.7.0",
    "redis[hiredis]>=5.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "sentence-transformers>=2.2.0",
    "python-dotenv>=1.2.1",
    "google-adk>=1.19.0",
    "langchain-google-genai>=3.2.0",
    "cachetools>=6.2.2",
]

[project.optional-dependencies]
langchain = [
    "langchain-core>=0.1.0",
    "langchain-google-genai>=0.0.5",
]
google-adk = [
    "google-adk>=0.1.0",
]
openai = [
    "openai>=1.0.0",
]
cohere = [
    "cohere>=5.0.0",
]
voyageai = [
    "voyageai>=0.1.0",
]
vertexai = [
    "google-cloud-aiplatform>=1.38.0",
]
observability = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
    "opentelemetry-exporter-jaeger>=1.20.0",
    "opentelemetry-instrumentation>=0.41b0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
all = [
    "semantic-cache[langchain,google-adk,openai,cohere,voyageai,vertexai,observability,dev]",
]

[project.urls]
Homepage = "https://github.com/vertector/semantic-cache"
Documentation = "https://github.com/vertector/semantic-cache#readme"
Repository = "https://github.com/vertector/semantic-cache"
Issues = "https://github.com/vertector/semantic-cache/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/semantic_cache"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312", "py313"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[dependency-groups]
dev = [
    "ipython>=8.31.0",
    "langchain-google-genai>=2.0.8",
    "python-dotenv>=1.0.1",
    "redisvl>=0.4.3",
]
